% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hegdata.R
\docType{class}
\name{hegdata}
\alias{hegdata}
\title{Data on Historical Ethnic Geography}
\description{
Provides an interface to retrieve the Historical Ethnic Geography dataset 
   published in Müller-Crepon, Schvitz and Cederman (Journal of Conflict Resolution, forthcoming).
}
\details{
For full information on the LEDA project and methodology,
    read the \href{https://github.com/carl-mc/hegdata/raw/master/docs/hegdata_paper.pdf}{paper}. 
    
   When using the \code{hegdata} package, please cite:
   Müller-Crepon, Carl, Guy Schvitz, and Lars-Erik Cederman (2023). 
   "Right-Peopling" the State: Nationalism, Historical Legacies, and Ethnic Cleansing in Europe, 1886-2020. 
   \emph{Journal of Conflict Resolution}, forthcoming.
}
\section{Usage}{

\preformatted{
# Initialize
heg.obj <- hegdata$new()

# Apply any LEDA method
# heg.obj$method() ## not run.
}
}

\examples{

# Initialize HEG object
heg.obj <- hegdata$new()

# Load raster for Polish people


## ------------------------------------------------
## Method `hegdata$new`
## ------------------------------------------------

library(hegdata)
heg.obj <- hegdata$new()


## ------------------------------------------------
## Method `hegdata$all_groups`
## ------------------------------------------------

library(hegdata)
heg.obj <- hegdata$new()
heg.obj$all_groups()


## ------------------------------------------------
## Method `hegdata$group_meta_data`
## ------------------------------------------------

library(hegdata)
heg.obj <- hegdata$new()
heg.obj$group_meta_data()


## ------------------------------------------------
## Method `hegdata$main_raster`
## ------------------------------------------------

library(hegdata)
heg.obj <- hegdata$new()
heg.obj$main_raster()


## ------------------------------------------------
## Method `hegdata$loadHEGGroup`
## ------------------------------------------------

library(hegdata)
heg.ob <- hegdata$new()
polish <- heg.obj$loadHEGGroup(group = "polish", year = 1918)
plot(polish)


## ------------------------------------------------
## Method `hegdata$loadHEGGroup_interpol`
## ------------------------------------------------

library(hegdata)
heg.ob <- hegdata$new()
polish <- heg.obj$loadHEGGroup_interpol(group = "polish", year = 1918)
plot(polish)

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-hegdata-new}{\code{hegdata$new()}}
\item \href{#method-hegdata-all_groups}{\code{hegdata$all_groups()}}
\item \href{#method-hegdata-group_meta_data}{\code{hegdata$group_meta_data()}}
\item \href{#method-hegdata-main_raster}{\code{hegdata$main_raster()}}
\item \href{#method-hegdata-loadHEGGroup}{\code{hegdata$loadHEGGroup()}}
\item \href{#method-hegdata-loadHEGGroup_interpol}{\code{hegdata$loadHEGGroup_interpol()}}
\item \href{#method-hegdata-clone}{\code{hegdata$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hegdata-new"></a>}}
\if{latex}{\out{\hypertarget{method-hegdata-new}{}}}
\subsection{Method \code{new()}}{
Initializes a new heegdata object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hegdata$new()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{library(hegdata)
heg.obj <- hegdata$new()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hegdata-all_groups"></a>}}
\if{latex}{\out{\hypertarget{method-hegdata-all_groups}{}}}
\subsection{Method \code{all_groups()}}{
Queries all group names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hegdata$all_groups()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Vector with all group names, alphabetically sorted
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{library(hegdata)
heg.obj <- hegdata$new()
heg.obj$all_groups()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hegdata-group_meta_data"></a>}}
\if{latex}{\out{\hypertarget{method-hegdata-group_meta_data}{}}}
\subsection{Method \code{group_meta_data()}}{
Queries all group meta data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hegdata$group_meta_data(smoothed = F)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{smoothed}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A dataset with one row per group-country-period combination. 
   The dataset contains the following variables:
   \code{gwcode} Gelditsch & Ward country code
   \code{fid} Unique ID per country-shape
   \code{group} Group name
   \code{start} Start year of period
   \code{end} End year of period
   \code{rowmin} Min row of this sub-raster in main raster frame (hegdata$main.ras)
   \code{rowmax} Max row of this sub-raster in main raster frame (hegdata$main.ras)
   \code{colmin} Min column of this sub-raster in main raster frame (hegdata$main.ras)
   \code{colmax} Max column of this sub-raster in main raster frame (hegdata$main.ras)
   \code{filename} File name of subraster
   \code{level} Level of group on language tree (org refers to original language node)
   \code{heg_main_match} Group name and language tree level concatenated
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{library(hegdata)
heg.obj <- hegdata$new()
heg.obj$group_meta_data()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hegdata-main_raster"></a>}}
\if{latex}{\out{\hypertarget{method-hegdata-main_raster}{}}}
\subsection{Method \code{main_raster()}}{
Returns the main raster frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hegdata$main_raster()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A spatial raster with value 0 for cells covered by HEG and NA for cells outside HEG's coverage.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{library(hegdata)
heg.obj <- hegdata$new()
heg.obj$main_raster()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hegdata-loadHEGGroup"></a>}}
\if{latex}{\out{\hypertarget{method-hegdata-loadHEGGroup}{}}}
\subsection{Method \code{loadHEGGroup()}}{
Loads raster for a group year using the baseline, non-smoothed version of the HEG data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hegdata$loadHEGGroup(group, year, verbose = F)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group}}{Specify the group name (lower or upper case, see $all_groups for all group names).}

\item{\code{year}}{Specify a year between 1886 and 2020}

\item{\code{verbose}}{Flag for level of verbosity. Defaults to FALSE.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns a raster of the extent of $main.ras (the full HEG frame) 
with values between 0 and 1 indicating the local population share 
belonging to the specified group as estimated in the HEG data.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{library(hegdata)
heg.ob <- hegdata$new()
polish <- heg.obj$loadHEGGroup(group = "polish", year = 1918)
plot(polish)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hegdata-loadHEGGroup_interpol"></a>}}
\if{latex}{\out{\hypertarget{method-hegdata-loadHEGGroup_interpol}{}}}
\subsection{Method \code{loadHEGGroup_interpol()}}{
Loads raster for a group year using the interpolated version of the HEG data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hegdata$loadHEGGroup_interpol(group, year, verbose = F)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group}}{Specify the group name (lower or upper case, see $all_groups for all group names).}

\item{\code{year}}{Specify a year between 1886 and 2020}

\item{\code{verbose}}{Flag for level of verbosity. Defaults to FALSE.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns a raster of the extent of $main.ras (the full HEG frame) 
with values between 0 and 1 indicating the local population share 
belonging to the specified group as estimated in the HEG data.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{library(hegdata)
heg.ob <- hegdata$new()
polish <- heg.obj$loadHEGGroup_interpol(group = "polish", year = 1918)
plot(polish)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hegdata-clone"></a>}}
\if{latex}{\out{\hypertarget{method-hegdata-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hegdata$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
